+++ CPUNeuralNetwork/main.cpp
@@ -11,17 +11,17 @@ int main(){
     dat.import_dataset(filename);
     cpu::Dataset x_train = dat.X_train_split();
     cpu::Dataset x_train_stand = x_train.standardizeMatrix();
-    std::vector<double> y_train = dat.y_train_split();
+    std::vector<float> y_train = dat.y_train_split();
 
     net.fit(x_train_stand, y_train);
 
     cpu::Dataset x_test = dat.X_test_split();
     cpu::Dataset x_test_stand = x_test.standardizeMatrix();
-    std::vector<double> y_test = dat.y_test_split();
+    std::vector<float> y_test = dat.y_test_split();
     dat.setValues(y_test);
-    double threeshold = 0.5;
-    std::vector<double> y_pred = net.perdict(x_test_stand, threeshold);
-    double acc = net.computeAccuracy(y_pred, y_test);
+    float threeshold = 0.5;
+    std::vector<float> y_pred = net.perdict(x_test_stand, threeshold);
+    float acc = net.computeAccuracy(y_pred, y_test);
     std::cout << acc << std::endl;
     
     cpu::Testing test;
